+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/CosmosAsyncScripts.java
@@ -8,7 +8,6 @@ import com.azure.cosmos.implementation.UserDefinedFunction;
 import com.azure.cosmos.model.CosmosAsyncStoredProcedureResponse;
 import com.azure.cosmos.model.CosmosAsyncTriggerResponse;
 import com.azure.cosmos.model.CosmosAsyncUserDefinedFunctionResponse;
-import com.azure.cosmos.model.CosmosPagedFlux;
 import com.azure.cosmos.model.CosmosStoredProcedureProperties;
 import com.azure.cosmos.model.CosmosStoredProcedureRequestOptions;
 import com.azure.cosmos.model.CosmosTriggerProperties;
@@ -90,7 +89,7 @@ public class CosmosAsyncScripts {
      * properties or an error.
      */
     public CosmosPagedFlux<CosmosStoredProcedureProperties> readAllStoredProcedures(FeedOptions options) {
-        return ModelBridgeInternal.createCosmosPagedFlux(pagedFluxOptions -> {
+        return new CosmosPagedFlux<>(pagedFluxOptions -> {
             setContinuationTokenAndMaxItemCount(pagedFluxOptions, options);
             return database.getDocClientWrapper()
                        .readStoredProcedures(container.getLink(), options)
@@ -135,7 +134,7 @@ public class CosmosAsyncScripts {
     public CosmosPagedFlux<CosmosStoredProcedureProperties> queryStoredProcedures(
         SqlQuerySpec querySpec,
         FeedOptions options) {
-        return ModelBridgeInternal.createCosmosPagedFlux(pagedFluxOptions -> {
+        return new CosmosPagedFlux<>(pagedFluxOptions -> {
             setContinuationTokenAndMaxItemCount(pagedFluxOptions, options);
             return database.getDocClientWrapper()
                        .queryStoredProcedures(container.getLink(), querySpec, options)
@@ -193,7 +192,7 @@ public class CosmosAsyncScripts {
      * error.
      */
     public CosmosPagedFlux<CosmosUserDefinedFunctionProperties> readAllUserDefinedFunctions(FeedOptions options) {
-        return ModelBridgeInternal.createCosmosPagedFlux(pagedFluxOptions -> {
+        return new CosmosPagedFlux<>(pagedFluxOptions -> {
             setContinuationTokenAndMaxItemCount(pagedFluxOptions, options);
             return database.getDocClientWrapper()
                        .readUserDefinedFunctions(container.getLink(), options)
@@ -240,7 +239,7 @@ public class CosmosAsyncScripts {
     public CosmosPagedFlux<CosmosUserDefinedFunctionProperties> queryUserDefinedFunctions(
         SqlQuerySpec querySpec,
         FeedOptions options) {
-        return ModelBridgeInternal.createCosmosPagedFlux(pagedFluxOptions -> {
+        return new CosmosPagedFlux<>(pagedFluxOptions -> {
             setContinuationTokenAndMaxItemCount(pagedFluxOptions, options);
             return database.getDocClientWrapper()
                        .queryUserDefinedFunctions(container.getLink(), querySpec, options)
@@ -295,7 +294,7 @@ public class CosmosAsyncScripts {
      * an error.
      */
     public CosmosPagedFlux<CosmosTriggerProperties> readAllTriggers(FeedOptions options) {
-        return ModelBridgeInternal.createCosmosPagedFlux(pagedFluxOptions -> {
+        return new CosmosPagedFlux<>(pagedFluxOptions -> {
             setContinuationTokenAndMaxItemCount(pagedFluxOptions, options);
             return database.getDocClientWrapper()
                        .readTriggers(container.getLink(), options)
@@ -336,7 +335,7 @@ public class CosmosAsyncScripts {
     public CosmosPagedFlux<CosmosTriggerProperties> queryTriggers(
         SqlQuerySpec querySpec,
         FeedOptions options) {
-        return ModelBridgeInternal.createCosmosPagedFlux(pagedFluxOptions -> {
+        return new CosmosPagedFlux<>(pagedFluxOptions -> {
             setContinuationTokenAndMaxItemCount(pagedFluxOptions, options);
             return database.getDocClientWrapper()
                        .queryTriggers(container.getLink(), querySpec, options)
