+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/CosmosPagedFlux.java
@@ -1,11 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.cosmos.model;
+package com.azure.cosmos;
 
 import com.azure.core.util.IterableStream;
 import com.azure.core.util.paging.ContinuablePagedFlux;
 import com.azure.cosmos.implementation.CosmosPagedFluxOptions;
+import com.azure.cosmos.model.FeedResponse;
 import reactor.core.CoreSubscriber;
 import reactor.core.publisher.Flux;
