+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/CosmosPagedIterable.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.cosmos.model;
+package com.azure.cosmos;
 
 import com.azure.core.util.paging.ContinuablePagedFlux;
 import com.azure.core.util.paging.ContinuablePagedIterable;
