+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/CosmosContainerTest.java
@@ -9,7 +9,6 @@ package com.azure.cosmos;
 import com.azure.cosmos.model.CosmosContainerProperties;
 import com.azure.cosmos.model.CosmosContainerRequestOptions;
 import com.azure.cosmos.model.CosmosContainerResponse;
-import com.azure.cosmos.model.CosmosPagedIterable;
 import com.azure.cosmos.model.FeedOptions;
 import com.azure.cosmos.model.IndexingMode;
 import com.azure.cosmos.model.IndexingPolicy;
