+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/CosmosTriggerTest.java
@@ -3,7 +3,6 @@
 
 package com.azure.cosmos;
 
-import com.azure.cosmos.model.CosmosPagedIterable;
 import com.azure.cosmos.model.CosmosTriggerProperties;
 import com.azure.cosmos.model.CosmosTriggerResponse;
 import com.azure.cosmos.model.FeedOptions;
