+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/CosmosUserTest.java
@@ -3,7 +3,6 @@
 
 package com.azure.cosmos;
 
-import com.azure.cosmos.model.CosmosPagedIterable;
 import com.azure.cosmos.model.CosmosUserProperties;
 import com.azure.cosmos.model.CosmosUserResponse;
 import com.azure.cosmos.model.FeedOptions;
