+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/rx/ReadFeedTriggersTest.java
@@ -5,7 +5,7 @@ package com.azure.cosmos.rx;
 import com.azure.cosmos.CosmosAsyncClient;
 import com.azure.cosmos.CosmosAsyncContainer;
 import com.azure.cosmos.CosmosClientBuilder;
-import com.azure.cosmos.model.CosmosPagedFlux;
+import com.azure.cosmos.CosmosPagedFlux;
 import com.azure.cosmos.model.CosmosTriggerProperties;
 import com.azure.cosmos.model.FeedOptions;
 import com.azure.cosmos.model.TriggerOperation;
