+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/rx/ReadFeedUsersTest.java
@@ -5,7 +5,7 @@ package com.azure.cosmos.rx;
 import com.azure.cosmos.CosmosAsyncClient;
 import com.azure.cosmos.CosmosAsyncDatabase;
 import com.azure.cosmos.CosmosClientBuilder;
-import com.azure.cosmos.model.CosmosPagedFlux;
+import com.azure.cosmos.CosmosPagedFlux;
 import com.azure.cosmos.CosmosDatabaseForTest;
 import com.azure.cosmos.model.CosmosUserProperties;
 import com.azure.cosmos.model.FeedOptions;
